pyproject.toml
[tool.poetry]
name = "Selenium Bot Python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
colorama = "^0.4.6"
discord = "^2.3.2"
requests = "^2.31.0"
selenium = "^4.12.0"
selenium-wire = "^5.1.0"
python-dotenv = "^1.0.0"
pyautogui = "^0.9.54"
mouseinfo = "^0.1.3"
webdriver-manager = "^4.0.0"
asynciolimiter = "^1.0.0b1"
selenium-async = "^0.1.0"

[tool.poetry.group.dev.dependencies]
asyncio = "^3.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
